# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-roundto",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `round_to(value, decimals)`.",
        "description": "This is a leaf: it delegates \"is this a number\" for `value` to\n`srvcs-isnumber` over HTTP (the single source of truth), then performs the\nrounding locally in f64. `decimals` is a small count read locally as an i64\n(>= 0). If the dependency is unreachable, this service reports itself\ndegraded rather than guessing.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundToResponse"
                }
              }
            }
          },
          "422": {
            "description": "value is not a number, or decimals is not a non-negative integer"
          },
          "500": {
            "description": "value passed validation but is not representable as a number"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value",
          "decimals"
        ],
        "properties": {
          "decimals": {
            "type": "object"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "RoundToResponse": {
        "type": "object",
        "required": [
          "value",
          "decimals",
          "result"
        ],
        "properties": {
          "decimals": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          },
          "value": {
            "type": "object"
          }
        }
      }
    }
  }
}